Fleshed out A LOT of the ability, character, status effect and SO system.
I'm currently attempting to make special scripts to hold ability functionality for easy ability customization.
The workflow consists of duplicating the current ability script Single_Target_Missle and then creating an asset for it, then removing the [CreateAssetMenu] attribute afterwards to avoid multiple copies.
Note: Nesting the SO object DIRECTLY inside the SO AbilityInformation caused serialization problems and random ability memory pointer problems. Storing the inner-most SO for Ability Behavior inside of a property of another class that is the property of the main SO seems to add a layer of seperation of sorts, allowing for correct serialization and no mixed up memory pointers. The AbilityBehavior class is now to be stored inside of GeneralAbilityInformation.

With this change, we should be able to utilize the targeting type enum inside this same data container for other purposes to assist with the AbilityBehavior logic.